Update main.c
Oh lord.
Well, I thought it was the pointer arithmetic, but it wasn't. So after 8 hours of slamming my head into a wall I eventually found out everything was falling apart because the array order[][] in main was being initialized before the global rows/cols were changed, causing it to access memory to assigned to the array. Tears were shed.